+++ packages/xo-server/src/xo-mixins/plugins.js
@@ -1,13 +1,39 @@
 import Ajv from 'ajv'
 import createLogger from '@xen-orchestra/log'
-
-import { PluginsMetadata } from '../models/plugin-metadata'
 import { invalidParameters, noSuchObject } from 'xo-common/api-errors'
+import { mapValues } from 'lodash'
+
+import replaceSensitiveValues from '../replace-sensitive-values'
+import { PluginsMetadata } from '../models/plugin-metadata'
 import { isFunction, mapToArray } from '../utils'
 
 // ===================================================================
 
 const log = createLogger('xo:xo-mixins:plugins')
+
+// this random value is used to obfuscate real data
+const OBFUSCATED_VALUE = 'c<,R"/|+9[-&|/pI!/}'
+
+const replaceObfuscatedValues = (newConfig, currentConfig) => {
+  if (newConfig === OBFUSCATED_VALUE) {
+    return currentConfig
+  }
+
+  let isArray
+
+  if (
+    newConfig === null ||
+    currentConfig === null ||
+    typeof newConfig !== 'object' ||
+    typeof currentConfig !== 'object' ||
+    (isArray = Array.isArray(newConfig)) !== Array.isArray(currentConfig)
+  ) {
+    return newConfig
+  }
+
+  const iteratee = (v, k) => replaceObfuscatedValues(v, currentConfig[k])
+  return isArray ? newConfig.map(iteratee) : mapValues(newConfig, iteratee)
+}
 
 export default class {
   constructor (xo) {
@@ -119,7 +145,7 @@ export default class {
       loaded,
       unloadable,
       version,
-      configuration,
+      configuration: replaceSensitiveValues(configuration, OBFUSCATED_VALUE),
       configurationPresets,
       configurationSchema,
       testable,
@@ -165,6 +191,14 @@ export default class {
   // save the new configuration.
   async configurePlugin (id, configuration) {
     const plugin = this._getRawPlugin(id)
+    const metadata = await this._getPluginMetadata()
+
+    if (metadata !== undefined) {
+      configuration = replaceObfuscatedValues(
+        configuration,
+        metadata.configuration
+      )
+    }
 
     await this._configurePlugin(plugin, configuration)